Pong: fingerprint: Sleep for 150ms on postEnroll
before enabling fod hbm.

Currently in our fingerprint implementation in kernel when the device exits hbm it causes the screen to get very dim for a moment or two, this is not noticeable when unlocking the device with a passcode but is definitely noticeable after a user completes fingeprint setup (After device setup in settings), so sleeping for sometime is a good hack as it avoids the screen getting too dim.

Signed-off-by: Abhay Singh Gill <abhaygill017@gmail.com>
Signed-off-by: N0m4n904 <noah.anleitner@halogenos.org>